requirements.txt
boto3>=1.34.0
psycopg2-binary>=2.9.9
PyPDF2>=3.0.0
langchain-text-splitters>=0.0.1
python-dotenv>=1.0.0
